\name{Ec_iAF1260_flux1.rda}

\docType{data}

\alias{Ec_iAF1260_flux1}

\title{E.coli model 1260 ORF's}

\description{This "list" form of the E.coli model iAF1260 
		was created from the S4 object created by readSBMLmod 
		from the package Sybil using the Sybil_2_FBA_obj function.}

\usage{Ec_iAF1260_flux1}

\format{A list containing the information required to perform FBA and annotations for intelligible output.}

\source{BiGG database}

\references{Feist AM, Henry CS, Reed JL, Krummenacker M, Joyce AR, Karp PD, Broadbelt LJ, Hatzimanikatis V and Palsson BO, A genome-scale metabolic reconstruction for Escherichia coli K-12 MG1655 that accounts for 1260 ORFs and thermodynamic information, Molecular Systems Biology.}
