\name{Ec_iJR904}
    
\docType{data}

\alias{Ec_iJR904}

\title{E.coli model 904 ORF's}

\description{This "list" form of the E.coli model iJR904 was 
created from the S4 object created by readSBMLmod from 
the package Sybil using the Sybil_2_FBA_obj function.}

\usage{Ec_iJR904}

\format{A list containing the information required to perform FBA and 
annotations for intelligible output.}

\source{BiGG database}

\references{Reed JL,Vo TD, Schilling CH and Palsson BO, An expanded 
genome-scale metabolic reconstruction for Escherichia coli K-12 
(iJR904 GSM/GPR) Genome Biology 2003}
