\name{Ecoli_core.rda}
\docType{data}
\alias{Ecoli_core}

\title{E.coli core model 72 reactions and 95 metabolites }

\description{This "list" form of the E.coli core model 
	was created from the S4 object created by readSBMLmod 
	from the package Sybil using the Sybil_2_FBA_obj function.}

\usage{Ecoli_core}

\format{A list containing the information required to perform FBA and 
annotations for intelligible output.}

\source{BiGG database}

\references{Systems Biology Properties of Reconstructed Networks}

\keyword{datasets}
