\name{H_pylori_iIT341}

    \docType{data}

	\alias{H_pylori_iIT341}

    \title{Helicobacter pylori by Ines Thiele}

    \description{This "list" form of the H. pylori model iIT341 was 
	created from the S4 object created by readSBMLmod from 
	the package Sybil using the Sybil_2_FBA_obj function.}

    \usage{H_pylori_iIT341}

    \format{A list containing the information required to perform FBA and annotations for intelligible output.}

    \source{BiGG database}

    \references{Thiele I, Vo TD, Price ND, Palsson BO, Expanded Metabolic Reconstruction of Helicobacter pylori (iIT341 GSM/GPR): an In Silico Genome-Scale Characterization of Single- and Double-Deletion Mutants, Journal of Bacteriology.}
