\name{H_sapien_Recon1}

\docType{data}

\alias{H_sapien_Recon1}

\title{H. sapien Recon1, human metabolic network reconstruction}

\description{This "list" form of the human model Recon1 was created 
from the S4 object created by readSBMLmod from the package Sybil 
using the Sybil_2_FBA_obj function.}

\usage{H_sapien_Recon1}

\format{A list containing the information required to perform FBA and 
annotations for intelligible output.}

\source{BiGG database}

\references{Duarte NC, Becker SA, Jamshidi N, Thiele I, Mo ML, Vo TD, Srivas R 
and Palsson BO, Global reconstruction of the human metabolic network based on 
genomic and bibliomic data, PNAS.}
