\name{M_barkeri_iAF629}

\docType{data}

\alias{M_barkeri_iAF629}

\title{Genome scale metabolic model for the archaeal methanogen M. Barkeri}

\description{This "list" form of the M.Barkeri model 
	was created from the S4 object created by readSBMLmod 
	from the package Sybil using the Sybil_2_FBA_obj function.}

\usage{M_barkeri_iAF629.rda}

\format{A list containing the information required to perform FBA and 
annotations for intelligible output.}

\source{BiGG database}

\references{Modeling methanogenesis with a genome scale metabolic 
reconstruction of Methanosarcina barkeri by Feist AM, Scholten JCM, Palsson BO}
