\name{M_tb_iNJ661}

    \docType{data}

    \alias{M_tb_iNJ661}

    \title{M. tuberculosis model iNJ661}

    \description{This "list" form of the M.tb model iNJ661 
	was created from the S4 object created by readSBMLmod 
	from the package Sybil using the Sybil_2_FBA_obj function.}

    \usage{M_tb_iNJ661}

    \format{A list containing the information required to perform FBA and annotations for intelligible output.}

    \source{BiGG database}

    \references{Bordbar A, Lewis NE, Schellenberger J, Palsson BO, Jamshidi N, Insight into human alveolar macrophage and M. tuberculosis interactions via metabolic reconstructions, Molecular Systems Biology.}
