\name{S_aureus_iSB619}

    \docType{data}

    \alias{S_aureus_iSB619}

    \title{Staphylococcus aureus model iSB619 by S.Becker}

    \description{This "list" form of the S. aureus model iSB619 
	was created from the S4 object created by readSBMLmod 
	from the package Sybil using the Sybil_2_FBA_obj function.}

    \usage{S_aureus_iSB619}

    \format{A list containing the information required to perform FBA and annotations for intelligible output.}

    \source{BiGG database}

    \references{Becker SA, Palsson BO, Genome-scale reconstruction of the metabolic network in Staphylococcus aureus N315: an initial draft to the two-dimensional annotation}
