\name{S_cerevisiae_IND750}

\docType{data}

\alias{S_cerevisiae_IND750}

\title{Saccharomyces Cerevisiae iND750}

\description{This "list" form of the yeast model iND750 was 
created from the S4 object created by readSBMLmod from 
the package Sybil using the Sybil_2_FBA_obj function.}

\usage{S_cerevisiae_IND750}

\format{A list containing the information required to perform FBA and 
annotations for intelligible output.}

\source{BiGG database}

\references{Duarte NC ,Herrgard MJ , Palsson BO, Reconstruction and Validation 
of Saccharomyces cerevisiae iND750, a Fully Compartmentalized Genome-Scale 
Metabolic Model, Genome Research}
