\name{Sybil_2_FBA_obj}
\alias{Sybil_S4_object}
\alias{SBML}
\alias{Sybil_2_FBA_obj}
\title{Sybil_2_FBA_obj, a function to convert a model generated by Sybil into the one used by abcdeFBA.}
\description{Changes the S4 object read by Sybil which is a non-standard R package into a list usable by abcdeFBA.}
\usage{
    Sybil_2_FBA_obj(Sybil_S4_object)
    }
\arguments{
	\item{Sybil_S4_object}{
	Is an S4 object of class modelorg}
}
\examples{
	#Model conversion
	#data(Ecoli_core) - # the Ecoli_core model included in the Sybil package
	#FBA_obj<-Sybil_2_FBA_obj(Ecoli_core)
}
